requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.1
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-multipart==0.0.6
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
aiofiles==23.2.1
mailjet-rest==1.3.4
cloudinary==1.41.0
